project/requirements.txt
robin_stocks
pandas
numpy
python-dotenv